requirements.txt
python-telegram-bot==13.7
pandas==1.3.5
python-dotenv==0.19.0
gspread==5.3.2  # Si usas Google Sheets
PyPDF2==2.0.0   # Para procesar PDFs